added Quotes and added Movie links for "video-prototyp-url"
Unfortunately, I did not record a Video on Werksschau. The Palakt was some times in the Corner fo my room and starting to mould, the tube was blocked and the elektrical transformer was dead. I cleaned everything and build a new tube but up to know I took only some Pictures and have not record a finished movie  ... sorry